:recycle: (TutorialTimeline) extract common steps
