+++ src/features/tutorial/components/profileTutorial/Timelines/TutorialTimelineSeventeen.tsx
@@ -3,12 +3,13 @@ import styled from 'styled-components/native'
 
 import { CreditComponentProps, CreditTimeline } from 'features/tutorial/components/CreditTimeline'
 import { EighteenBlockDescription } from 'features/tutorial/components/profileTutorial/EighteenBlockDescription'
-import { InformationStepContent } from 'features/tutorial/components/profileTutorial/InformationStepContent'
+import {
+  eighteenSeparatorStep,
+  lastStep,
+} from 'features/tutorial/components/profileTutorial/Timelines/CommonSteps'
 import { UnderageBlockDescription } from 'features/tutorial/components/profileTutorial/UnderageBlockDescription'
 import { TutorialTypes } from 'features/tutorial/enums'
 import { useDepositAmountsByAge } from 'shared/user/useDepositAmountsByAge'
-import { Warning } from 'ui/svg/icons/BicolorWarning'
-import { Offers } from 'ui/svg/icons/Offers'
 import { Spacer, Typo } from 'ui/theme'
 
 interface Props {
@@ -59,14 +60,6 @@ export const TutorialTimelineSeventeen = ({ activatedAt }: Props) => {
   )
 }
 
-const GreyOffers = styled(Offers).attrs(({ theme }) => ({
-  color: theme.colors.greySemiDark,
-}))``
-
-const GreyWarning = styled(Warning).attrs(({ theme }) => ({
-  color: theme.colors.greySemiDark,
-}))``
-
 const StyledCaption = styled(Typo.Caption)(({ theme }) => ({
   color: theme.colors.greyDark,
 }))
@@ -83,22 +76,7 @@ const PastStepBlock = ({ text }: { text: string }) => {
 
 const SeventeenStepperProps: CreditComponentProps[] = [
   { creditStep: 17, children: <UnderageBlockDescription /> },
-  {
-    creditStep: 'information',
-    iconComponent: <GreyWarning />,
-    children: (
-      <InformationStepContent title="La veille de tes 18 ans" subtitle="Ton crédit est remis à 0" />
-    ),
-  },
+  eighteenSeparatorStep,
   { creditStep: 18, children: <EighteenBlockDescription /> },
-  {
-    creditStep: 'information',
-    iconComponent: <GreyOffers />,
-    children: (
-      <InformationStepContent
-        title="Au bout de 2 ans, ton crédit expire… Mais l’aventure continue&nbsp;!"
-        subtitle="Tu peux continuer à réserver des offres gratuites autour de chez toi."
-      />
-    ),
-  },
+  lastStep,
 ]
